move client declarations
all client declarations are now in `~/clients.ts`
