--- server/src/lib/db.ts
@@ -1,8 +1,0 @@
-import Knex from 'knex';
-
-const db = Knex({
-  client: 'pg',
-  connection: process.env.DATABASE_URL,
-});
-
-export default db;
