--- server/src/lib/redis.ts
@@ -1,8 +1,0 @@
-import { createClient } from 'redis';
-
-const redis = createClient({
-  url: process.env.REDIS_URL,
-  password: process.env.REDIS_PASSWORD,
-});
-
-export default redis;
