+++ server/src/routes/me/tasks.ts
@@ -1,5 +1,5 @@
 import { Router } from 'express';
-import db from '~/lib/db';
+import { db } from '~/clients';
 import authenticate from '~/middlewares/authenticate';
 
 const router = Router();
@@ -11,7 +11,7 @@ router.get('/', authenticate, (req, res) => {
     .where({ 'projects.user_id': req.accessToken.user_id })
     .then(tasks => {
       res.status(200).json({
-        success: true, 
+        success: true,
         data: { tasks },
       });
     })
